requirements.txt
chainlit
langchain
langchain_community
PyPDF2
chromadb
groq
langchain-groq
ollama
python-dotenv
InstructorEmbedding